Move flash helper into own module and add specs
