+++ app/helpers/application_helper.rb
@@ -4,13 +4,4 @@ module ApplicationHelper
     page_title!(t('.title', label: object))
   end
 
-  def flash_class(level)
-    case level.to_sym
-    when :notice then "positive message"
-    when :success then "positive message"
-    when :error then "error message"
-    when :alert then "negative message"
-    end
-  end
-
 end
